# Performance improvement during run of instrumented code