@tarekgh 2018-08-13
@tonerdo I thought the same when I worked previously on the perf issue. I didn't spend time during then to investigate more but what you have said align what I was thinking too.