@tarekgh 2018-08-22
@pjanotti  I have added a comment but I am not sure if you answered it my question. why we introduced StringHashSuffixComparer  and didn't use StringComparer.Ordinal?