@tarekgh 2018-08-22
The hashing method can produce deceiving results and we may be lucky so far.
strings like:

"a" and "aaaaaaaaa" will produce the same hash code. if we are planning to change this anyway in the near future, then that is ok. otherwise, this may cause unexpected issues and will be hard to investigate.