@pjanotti 2018-08-22
> "a" and "aaaaaaaaa" will produce the same hash code.

Hashes in this case will be different. The "known" case of hash collision introduced by this PR is if the strings have the same length and the last 8 chars are the same, of course there could be the normal "statistical collisions". That said if there are a very high number of the collisions the performance will suffer but correctness won't be affected.